requirements.txt
coverage>=5.5, <6.0 #6.0 remove the support of python 3.6
Django>=3.2, <4.0
django-cors-headers>=3.7, <3.11 #3.11 remove the support of python3.6
django-ipware>=4.0, <5.0
flake8>=4.0, <5.0
importlib-metadata>=4.0, <4.3;python_version<"3.8" #Add in the standard library in 3.8. Flake8 doesn't handle more than 4.3
marshmallow-jsonapi>=0.24.0, <1.0
oic>=1.3, <2.0 # the max available python version is 3.9
psycopg2>=2.9, <3.0
pytest>=6.2, <7
pytest-cov>=3.0, <4.0
pytest-django>=4.5, <5.0
pytest-pythonpath>=0.7, <1.0
python-dateutil>=2.8, <3.0
python-jose>=3.3, <4.0
requests>=2.25, <3.0
marshmallow>=3.14,<3.15 #3.15 remove the support of python3.6
tox>=3.24, < 4.0
tox-pyenv>=1.1,<2.0
pre-commit>=2.17,<3.0
backports.zoneinfo>=0.2,<1.0;python_version<"3.9" #Add in the standard library in 3.9.
pandas>=1.1,<=1.5
freezegun>=1.1, <2.0
